Clone SVUnit as a source dependency
It was kind of tricky to get working. Mostly this is because SVUnit itself
isn't built using Gradle.

The easy part is declaring the dependency and that it is a source dependency.
To not pollute the compile, we declare it as a dependency to a new
configuration. We also declare a task that depends on this configuration,
which, when executed, will trigger the clone of the Git repo.

Since SVUnit isn't built with Gradle, we have to inject a plugin in it. We do
this with the extra included build that contains the 'svunit-build' plugin.
This plugin will programmatically create a Gradle project with the required
properties. It needs to get the name, group and version right, so that
dependency substitution works. it needs to define an outgoing configuration and
an artifact linked to it. The dependency resolution is done using the
'cloneSvunit' attribute.

It might be the case that this is too much code. I'll try to simplify it a bit
more.