Do away with setup script for unit testing
Loading SVUnit can be done in the run script itself.

This fixes any issues that might come up when upgrading to a newer version of
SVUnit. Remember that the version of SVUnit is present in the directory path
where it gets cloned. Let's assume that we have a version of SVUnit loaded, we
upgrade the version, then try to load it again. SVunit will ignore the load of
the new version, because it will see its environment variables and assume it's
already loaded. Trying to run would lead to errors, because now we won't have
'runSVUnit' in 'PATH'.