Add ability to get value of a string variable
Note that there is no need to also test this for _static_ string variables (as we did for `int`). While static variables require special handling, this is not specific to the type of the variable. Writing the tests for static `int` already forced us to implement this special handling for static variables.