Trigger error when trying to get value from non-string array handles
The currentl structure (to check whether the array is of type string inside the function to get the value from a string array) will not work when we implement support for other types of arrays. Then the case statement inside the `get` function will have to check of what type the array is.